+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [1.4.6] - 2026-01-19
+### Changed
+- Update package dependencies. [#46647]
+
 ## [1.4.5] - 2026-01-14
 ### Changed
 - Internal updates.
@@ -1612,6 +1616,7 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[1.4.6]: https://github.com/Automattic/jetpack-components/compare/1.4.5...1.4.6
 [1.4.5]: https://github.com/Automattic/jetpack-components/compare/1.4.4...1.4.5
 [1.4.4]: https://github.com/Automattic/jetpack-components/compare/1.4.3...1.4.4
 [1.4.3]: https://github.com/Automattic/jetpack-components/compare/1.4.2...1.4.3