+++ projects/js-packages/idc/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA IDC package releases.
 
+## 1.0.43 - 2026-01-19
+### Changed
+- Update package dependencies. [#46647]
+
 ## 1.0.42 - 2026-01-14
 ### Changed
 - Update dependencies. [#46035]