+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.7.19 - 2026-01-19
+### Changed
+- Remove unused `wp-theme-slug` fallback. It can be added back (everywhere in monorepo tooling) if we ever do have themes in the monorepo. [#46601]
+- Use `beta-plugin-slug` as a fallback for plugin textdomains if `wp-plugin-slug` isn't set. [#46601]
+- Update package dependencies. [#46647]
+
 ## 3.7.18 - 2026-01-07
 ### Changed
 - Update package dependencies. [#46456]