+++ projects/packages/newsletter/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-01-19
+### Added
+- Implement bits of infrastructure for newsletter settings UI. [#46470]
+
+### Changed
+- Update package dependencies. [#46552] [#46647]
+
 ## [0.1.4] - 2026-01-12
 ### Changed
 - Update package dependencies. [#46456]
@@ -28,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update package dependencies. [#46143]
 
+[0.2.0]: https://github.com/Automattic/jetpack-newsletter/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/Automattic/jetpack-newsletter/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/Automattic/jetpack-newsletter/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Automattic/jetpack-newsletter/compare/v0.1.1...v0.1.2