+++ projects/packages/publicize/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.71.0] - 2026-01-19
+### Added
+- Add per-network customization with global toggle and per-connection message/media overrides. [#46563]
+
+### Changed
+- Disable customization form when connection is disabled. [#46588]
+- Update package dependencies. [#46647]
+
 ## [0.70.0] - 2026-01-14
 ### Changed
 - Decouple `SharePostForm` component from the store by adding optional props for controlled mode. [#46545]
@@ -1207,6 +1215,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies.
 - Update package.json metadata.
 
+[0.71.0]: https://github.com/Automattic/jetpack-publicize/compare/v0.70.0...v0.71.0
 [0.70.0]: https://github.com/Automattic/jetpack-publicize/compare/v0.69.2...v0.70.0
 [0.69.2]: https://github.com/Automattic/jetpack-publicize/compare/v0.69.1...v0.69.2
 [0.69.1]: https://github.com/Automattic/jetpack-publicize/compare/v0.69.0...v0.69.1